# fix: use correct data source for querying max capacity on arm64-macOS