@WenqingZong 2023-05-22
For more discussion, please see: https://github.com/ClementTsang/bottom/issues/1164